refactor: `em` –> `animated-underline`
