+++ components/about/vision.tsx
@@ -1,7 +1,7 @@
 import AnimatedText from "../animated-text";
 import AnimatedObject from "../animated-object";
-import { motion } from "framer-motion";
 import Image from "next/image";
+import AnimatedUnderline from "../animated-underline";
 
 export default function Vision() {
   return (
@@ -16,14 +16,9 @@ export default function Vision() {
           </h1>
           <p className="mb-6 text-3xl font-medium">
             Uplifting communities through{" "}
-            <span className="relative z-0">
-              <motion.span
-                whileInView={{ width: ["0%", "100%"] }}
-                className="absolute left-4 bottom-0 -z-10 h-3 origin-left bg-amber-300/50"
-                transition={{ duration: 1.5 }}
-              />
+            <AnimatedUnderline className="left-4 bottom-0 bg-amber-300/50">
               technology
-            </span>
+            </AnimatedUnderline>
           </p>
           <p className="mb-8 text-base">
             Our organization envisions itself as a community of tech enthusiasts
