+++ components/team/comms-team.tsx
@@ -1,4 +1,4 @@
-import AnimatedEm from "../animated-em";
+import AnimatedUnderline from "../animated-underline";
 import ParallaxObject from "../parallax-object";
 import extractPortraits from "../../lib/extractPortraits";
 import { MemberInCMS } from "../../types/member-in-cms";
@@ -15,9 +15,9 @@ export default function CommsTeam({ allCommsData }: Props) {
     >
       <div className="max-w-sm px-10 lg:order-2 xl:max-w-md">
         <h2>
-          <AnimatedEm emClassName="bg-blue-400/50 bottom-0 left-4">
+          <AnimatedUnderline className="bottom-0 left-4 bg-blue-400/50">
             Communications
-          </AnimatedEm>
+          </AnimatedUnderline>
         </h2>
         <p className="mt-4">
           The Communications Department is in charge of crafting promotional
